+ move method MainCrawlWorker.SyncCrawlIntervalWithConfig() and field _timer into a new abstract class CyclicCrawlWorker to share them with RetryCrawlWorker * no longer delay the start of initial Retry() @ RetryCrawlWorker.ExecuteAsync() * now will eager initial all keyed CrawlerLocks singleton instances, in order to sync its timer of WithLogTrace @ MainCrawlWorker.ctor()
* mark the return type of method GetPageFromResponse() as nullable @ BaseCrawler.cs
* update appsettings.json with more proper default values
* only copy config files when they are modified @ tbm.Crawler.csproj
@ crawler

+++ crawler/src/Tieba/ClientRequesterTcs.cs
@@ -45,7 +45,7 @@ namespace tbm.Crawler
         {
             if (!ShouldLogTrace()) return;
             _logger.LogTrace("TCS: queueLen={} maxLimitRps={:F2} avgRps={:F2} elapsed={:F1}s",
-                QueueLength, MaxRps, AverageRps, (float)_stopwatch.ElapsedMilliseconds / 1000);
+                QueueLength, MaxRps, AverageRps, _stopwatch.ElapsedMilliseconds / 1000f);
             if (_config.GetValue("LogTrace:ResetAfterLog", false)) ResetAverageRps();
         }
